move generate methods to the init traits
